# maxBytesPerRead channel configuration