@Scottmitch 2015-06-12
@trustin - Would you mind reviewing this one?