@trustin 2015-06-30
Do the new configuration properties work only when the `RecvByteBufAllocator` is aware of them? If so, why don't we move them to the `RecvByteBufAllocator` implementation?