@Scottmitch 2015-06-30
Yes. Currently `MaxBytesRecvByteBufAllocator` is the only consumer of the new properties.  The `RecvByteBufAllocator` implementations have also encapsulated the responsibility of the existing `getMaxMessagesPerRead()`.  I think if we move the new parameters it may make sense to also move `getMaxMessagesPerRead()` into the respective implementation that uses it (this change is only for 4.1 and 5.0).  WDYT?