@Scottmitch 2015-06-30
@trustin - Another thing to consider is moving these configuration options from `ChannelConfig` to an implementation of `RecvByteBufAllocator` may make it a bit more cumbersome to change on the fly. For example this may require conditionals based upon the run time type of the `RecvByteBufAllocator`. For example:

```java
RecvByteBufAllocator recvAlloc = channel.config().getRecvByteBufAllocator();
if (recvAlloc instanceof MaxBytesRecvByteBufAllocator) {
  MaxBytesRecvByteBufAllocator maxRecvAlloc = (MaxBytesRecvByteBufAllocator) recvAlloc;
  maxRecvAlloc.setMaxMessagesPerRead(...);
  maxRecvAlloc.setMaxBytesPerIndividualRead(...);
} else if (recvAlloc instanceof ...) {
}
```