@trustin 2015-07-01
> I think if we move the new parameters it may make sense to also move `getMaxMessagesPerRead()` into the respective implementation that uses it (this change is only for 4.1 and 5.0). WDYT?

Sounds good to me!

> Another thing to consider is moving these configuration options from `ChannelConfig` to an implementation of `RecvByteBufAllocator` may make it a bit more cumbersome to change on the fly.

Yeah, how about using the type parameter for the return type for loose typing to lessen the burden?

```java
interface ChannelConfig {
    <T extends RecvByteBufAllocator> T getRecvByteBufAllocator();
}

MaxBytesRecvByteBufAllocator alloc = ch.config().getRecvByteBufAllocator();
...
```

This won't work when a user is not sure about the type of the current `RecvByteBufAllocator`, though...