@Scottmitch 2015-07-01
@trustin - Thanks for the feedback. I will move forward with these updates and ping you when ready.