@trustin 2015-07-07
@Scottmitch LGTM. @normanmaurer, any comments?