@normanmaurer 2015-07-07
@Scottmitch a few more comments and questions. Good work!