@Scottmitch 2015-07-16
@trustin - I would like to merge this but there are just a few followup discussion items above.  Would you mind weighing in?