@Scottmitch 2015-08-06
@trustin - I guess now that this doesn't break any APIs we could also consider putting this in 4.0. It would be nice to be consistent between 4.0 and 4.1...WDYT?