@Scottmitch 2015-08-06
Ran some quick perf tests on the before and after for this PR just to make sure there was no regression, and things look good (using NIO):

before
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    22.07ms   33.58ms   1.11s    85.72%
    Req/Sec   132.23k    33.88k  241.31k    66.05%
  251377804 requests in 2.00m, 33.24GB read
Requests/sec: 2093223.99
Transfer/sec:    283.47MB
```

this PR
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    21.36ms   34.33ms   1.28s    85.67%
    Req/Sec   136.39k    29.18k  231.36k    67.75%
  259304381 requests in 2.00m, 34.29GB read
Requests/sec: 2159239.24
Transfer/sec:    292.41MB
```