@Scottmitch 2015-08-06
Cherry-picked 4.1 (cf171ff) master (a48e5c7)